Update doc_standard.md
Change to the workflow to make it a little lighter.

I suggest that for documentation changes we generally are going to commit directly to the master branch, rather than initiating pull requests. Branching is absolutely essential for code, where there are usually complex and difficult to understand implications and a single wrong keystroke often spells failure. For documentation, its probably better to err on the side of speed. So I suggest we lighten the process, we can focus on reversion if we don't like the proposed change.