Update doc_standard.md
DO we really need a separate directory for pages that have no or only one image. The majority of our pages are text only. We also organize everything in chapters with 5-7 pages per chapter. Would it not make more sense just to add all the images of every page in one chapter. What do you think?