Update site_survey.md
I disagree with the tone of the original document. The easiest way to proceed is to have a dedicated 16.04 system on an older dedicated laptop, not installing virtualization software for Mac or Windows systems. Insisting on virtualization is adding a layer of complexity that is likely to add frustration to an already.  An easy way to get started would be to get an SD card, a Pi3, keyboard, mouse and HDMI monitor and start running simulations before getting a robot!